fix: get token from resource to avoid duplicate warnings
